requirements.txt
Django==3.2.5
psycopg2-binary==2.9.1
djangorestframework==3.12.4
django-cors-headers==3.11.0
PyJWT==2.3.0

